exporter: exporter dir and http_server.h
* move all related files to the exporter to src/exporter
* move http_server logic to a new file under src/exporter/http_server.*
