+++ src/exporter/http_server.cc
@@ -1,0 +1,210 @@
+#include "http_server.h"
+#include "exporter/DaemonMetricCollector.h"
+#include "common/hostname.h"
+
+#include <iostream>
+#include <map>
+#include <boost/beast/core.hpp>
+#include <boost/beast/http.hpp>
+#include <boost/beast/version.hpp>
+#include <boost/asio.hpp>
+#include <chrono>
+#include <cstdlib>
+#include <ctime>
+#include <iostream>
+#include <memory>
+#include <string>
+#include <boost/thread/thread.hpp>
+
+namespace beast = boost::beast;         // from <boost/beast.hpp>
+namespace http = beast::http;           // from <boost/beast/http.hpp>
+namespace net = boost::asio;            // from <boost/asio.hpp>
+using tcp = boost::asio::ip::tcp;       // from <boost/asio/ip/tcp.hpp>
+
+namespace my_program_state
+{
+    std::size_t
+    request_count()
+    {
+        static std::size_t count = 0;
+        return ++count;
+    }
+
+    std::time_t
+    now()
+    {
+        return std::time(0);
+    }
+}
+
+class http_connection : public std::enable_shared_from_this<http_connection>
+{
+public:
+    http_connection(tcp::socket socket)
+        : socket_(std::move(socket))
+    {
+    }
+
+    // Initiate the asynchronous operations associated with the connection.
+    void start()
+    {
+        read_request();
+        check_deadline();
+    }
+
+private:
+    tcp::socket socket_;
+    beast::flat_buffer buffer_{8192};
+    http::request<http::dynamic_body> request_;
+    http::response<http::dynamic_body> response_;
+
+    net::steady_timer deadline_{
+        socket_.get_executor(), std::chrono::seconds(60)};
+
+    // Asynchronously receive a complete request message.
+    void read_request()
+    {
+        // auto self = shared_from_this();
+
+        http::async_read(
+            socket_,
+            buffer_,
+            request_,
+            [&](beast::error_code ec,
+                std::size_t bytes_transferred)
+            {
+                boost::ignore_unused(bytes_transferred);
+                if(!ec)
+                    process_request();
+            });
+    }
+
+    // Determine what needs to be done with the request message.
+    void process_request()
+    {
+        response_.version(request_.version());
+        response_.keep_alive(false);
+
+        switch(request_.method())
+        {
+        case http::verb::get:
+            response_.result(http::status::ok);
+            response_.set(http::field::server, "Beast");
+            create_response();
+            break;
+
+        default:
+            // We return responses indicating an error if
+            // we do not recognize the request method.
+            response_.result(http::status::bad_request);
+            response_.set(http::field::content_type, "text/plain");
+            beast::ostream(response_.body())
+                << "Invalid request-method '"
+                << std::string(request_.method_string())
+                << "'";
+            break;
+        }
+
+        write_response();
+    }
+
+    // Construct a response message based on the program state.
+    void create_response()
+    {
+        std::cout << "Got request on " << request_.target() << std::endl;
+        if(request_.target() == "/metrics")
+        {
+            response_.set(http::field::content_type, "text/plain");
+            DaemonMetricCollector &collector = collector_instance();
+            std::string metrics = collector.get_metrics();
+            beast::ostream(response_.body()) << "Perf Counters\n" << metrics << std::endl;
+        }
+        else
+        {
+            response_.result(http::status::not_found);
+            response_.set(http::field::content_type, "text/plain");
+            beast::ostream(response_.body()) << "File not found\r\n";
+        }
+    }
+
+    // Asynchronously transmit the response message.
+    void write_response()
+    {
+        auto self = shared_from_this();
+
+        response_.content_length(response_.body().size());
+
+        http::async_write(
+            socket_,
+            response_,
+            [self](beast::error_code ec, std::size_t)
+            {
+                self->socket_.shutdown(tcp::socket::shutdown_send, ec);
+                self->deadline_.cancel();
+            });
+    }
+
+    // Check whether we have spent enough time on this connection.
+    void check_deadline()
+    {
+        auto self = shared_from_this();
+
+        deadline_.async_wait(
+            [self](beast::error_code ec)
+            {
+                if(!ec)
+                {
+                    // Close socket to cancel any outstanding operation.
+                    self->socket_.close(ec);
+                }
+            });
+    }
+};
+
+// "Loop" forever accepting new connections.
+void http_server(tcp::acceptor& acceptor, tcp::socket& socket)
+{
+  acceptor.async_accept(socket,
+      [&](beast::error_code ec)
+      {
+          std::cout << "async accept" << std::endl;
+          if(!ec)
+              std::make_shared<http_connection>(std::move(socket))->start();
+          http_server(acceptor, socket);
+      });
+}
+
+std::string dns_lookup(std::string hostname) {
+    boost::asio::io_service io_service;
+    boost::asio::ip::tcp::resolver resolver(io_service);
+    boost::asio::ip::tcp::resolver::query query(hostname, "9085");
+    boost::asio::ip::tcp::resolver::iterator iter = resolver.resolve(query);
+    boost::asio::ip::tcp::endpoint endpoint = iter->endpoint();
+    std::string ip_address = endpoint.address().to_string();
+
+    return ip_address;
+}
+
+void http_server_thread_entrypoint() {
+  try
+    {
+        std::string hostname = ceph_get_short_hostname();
+
+        std::string ip_address = dns_lookup(hostname);
+        auto const address = net::ip::make_address(ip_address);
+        unsigned short port = static_cast<unsigned short>(std::atoi("9085"));
+
+        net::io_context ioc{1};
+
+        tcp::acceptor acceptor{ioc, {address, port}};
+        tcp::socket socket{ioc};
+        http_server(acceptor, socket);
+        std::cout << "Http server started" << std::endl;
+        ioc.run();
+    }
+    catch(std::exception const& e)
+    {
+        std::cerr << "Error: " << e.what() << std::endl;
+        exit(EXIT_FAILURE);
+    }
+}
