+++ src/exporter/http_server.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void http_server_thread_entrypoint();
